Create CoronaQuestion.kt
Created a class for a Question object which compartmentalizes all the fields that a question will require through its lifespan. References translator methods, scrapers, nlp's. When ready to display all the information desired for the user,, all the data for the question is in one unit and ready to be extracted.