Update deduplicate_with_sentenceBert.py
Added a Question aggregate, which compartmentalizes all the fields that a question will require through its lifespan. References translator methods, scrapers, nlp's and its embeddings (not explicitly coded). When ready to display all the information desired for user viewing, all the data for the question is in one unit and ready to be extracted.