Update eval.py to include Composite Pattern
Added a Composite pattern so that we can treat a group of questions and a singular question uniformly, while attempting to match user input to questions in the database. The idea is to iterate through the composition which is organized based on relations computed by the eval modules, matching user input to diverging compositions until we reached the closest matched question.